Creado un nuevo componente para evitar duplicidad
